Speed up YNAB and ensure the plugin stops checking when it disappears
By speeding up the YNAB checks to 20 seconds,
it also uncovered a bug in the original implementation,
wherein the `quit` channel was never being called.

If you had the YNAB plugin on the Stream Deck,
then navigated to another profile for only 1 or 2 minutes,
before navigating back to the YNAB plugin,
it would spawn a second instance of the plugin,
and both instances would be checking every X seconds.

This resulted in the rate limit being breached.